refactor: change v2 api dir and user new manager method issue #1361
