+++ src/api_server/app/server.go
@@ -20,7 +20,6 @@ import (
 	"github.com/emicklei/go-restful"
 
 	"configcenter/src/api_server/app/options"
-	"configcenter/src/api_server/ccapi/logics/v2"
 	apisvc "configcenter/src/api_server/service"
 	"configcenter/src/api_server/service/v3"
 	"configcenter/src/apimachinery"
@@ -50,7 +49,7 @@ func Run(ctx context.Context, op *options.ServerOption) error {
 		return fmt.Errorf("new api machinery failed, err: %v", err)
 	}
 
-	v2Service := new(apisvc.Service)
+	v2Service := apisvc.NewService()
 	v3Service := new(v3.Service)
 	v3Service.Client, err = util.NewClient(&util.TLSClientConfig{})
 	if err != nil {
@@ -64,7 +63,7 @@ func Run(ctx context.Context, op *options.ServerOption) error {
 
 	ctnr := restful.NewContainer()
 	ctnr.Router(restful.CurlyRouter{})
-	ctnr.Add(v2Service.V2WebService())
+	ctnr.Add(v2Service.WebService())
 	ctnr.Add(v3Service.V3WebService())
 	ctnr.Add(v3Service.V3Healthz())
 	server := backbone.Server{
@@ -93,20 +92,15 @@ func Run(ctx context.Context, op *options.ServerOption) error {
 		return fmt.Errorf("new backbone failed, err: %v", err)
 	}
 
-	v2Service.Engine = engine
-	v2Service.Logics = &logics.Logics{Engine: engine}
+	v2Service.SetEngine(engine)
 	v3Service.Engine = engine
 	apiSvr.Core = engine
-	apiSvr.Service = v2Service
-	apiSvr.Logic = v2Service.Logics
 	select {}
 	return nil
 }
 
 type APIServer struct {
-	Core    *backbone.Engine
-	Service *apisvc.Service
-	Logic   *logics.Logics
+	Core *backbone.Engine
 }
 
 func (h *APIServer) onHostConfigUpdate(previous, current cc.ProcessConfig) {
