+++ src/api_server/service/v2/addhost.go
@@ -10,13 +10,13 @@
  * limitations under the License.
  */
 
-package service
+package v2
 
 import (
 	"context"
 	"strings"
 
-	"configcenter/src/api_server/ccapi/logics/v2/common/converter"
+	"configcenter/src/api_server/logics/v2/common/converter"
 	"configcenter/src/common"
 	"configcenter/src/common/blog"
 	"configcenter/src/common/metadata"
