+++ src/api_server/service/v2/gse.go
@@ -10,18 +10,18 @@
  * limitations under the License.
  */
 
-package service
+package v2
 
 import (
 	"context"
 
-	"configcenter/src/api_server/ccapi/logics/v2/common/converter"
-	"configcenter/src/api_server/ccapi/logics/v2/common/utils"
+	"github.com/emicklei/go-restful"
+
+	"configcenter/src/api_server/logics/v2/common/converter"
+	"configcenter/src/api_server/logics/v2/common/utils"
 	"configcenter/src/common"
 	"configcenter/src/common/blog"
 	"configcenter/src/common/util"
-
-	"github.com/emicklei/go-restful"
 )
 
 func (s *Service) getAgentStatus(req *restful.Request, resp *restful.Response) {
