+++ src/api_server/service/v2/property.go
@@ -10,14 +10,14 @@
  * limitations under the License.
  */
 
-package service
+package v2
 
 import (
 	"context"
 
-	"configcenter/src/api_server/ccapi/logics/v2/common/converter"
-	"configcenter/src/api_server/ccapi/logics/v2/common/defs"
-	"configcenter/src/api_server/ccapi/logics/v2/common/utils"
+	"configcenter/src/api_server/logics/v2/common/converter"
+	"configcenter/src/api_server/logics/v2/common/defs"
+	"configcenter/src/api_server/logics/v2/common/utils"
 	"configcenter/src/common"
 	"configcenter/src/common/blog"
 	"configcenter/src/common/util"
