+++ src/api_server/service/v2/set.go
@@ -10,7 +10,7 @@
  * limitations under the License.
  */
 
-package service
+package v2
 
 import (
 	"context"
@@ -20,8 +20,8 @@ import (
 	"strconv"
 	"strings"
 
-	"configcenter/src/api_server/ccapi/logics/v2/common/converter"
-	"configcenter/src/api_server/ccapi/logics/v2/common/utils"
+	"configcenter/src/api_server/logics/v2/common/converter"
+	"configcenter/src/api_server/logics/v2/common/utils"
 	"configcenter/src/common"
 	"configcenter/src/common/blog"
 	"configcenter/src/common/mapstr"
