+++ src/api_server/service/v2/userapi.go
@@ -10,7 +10,7 @@
  * limitations under the License.
  */
 
-package service
+package v2
 
 import (
 	"context"
@@ -19,7 +19,7 @@ import (
 	"strconv"
 	"strings"
 
-	"configcenter/src/api_server/ccapi/logics/v2/common/converter"
+	"configcenter/src/api_server/logics/v2/common/converter"
 	"configcenter/src/common"
 	"configcenter/src/common/blog"
 	"configcenter/src/common/mapstr"
